+++ litebox_platform_lvbs/src/lib.rs
@@ -389,6 +389,7 @@ pub struct LinuxKernel<Host: HostInterface> {
     host_and_task: core::marker::PhantomData<Host>,
     page_table_manager: PageTableManager,
     vtl1_phys_frame_range: PhysFrameRange<Size4KiB>,
+    end_of_boot: core::sync::atomic::AtomicBool,
 }
 
 /// [`litebox::platform::common_providers::userspace_pointers::ValidateAccess`]
@@ -621,7 +622,19 @@ impl<Host: HostInterface> LinuxKernel<Host> {
             host_and_task: core::marker::PhantomData,
             page_table_manager: PageTableManager::new(base_pt),
             vtl1_phys_frame_range: vtl1_range,
+            end_of_boot: core::sync::atomic::AtomicBool::new(false),
         }))
+    }
+
+    /// Whether VTL1's window of trusting VTL0 has closed.
+    pub(crate) fn end_of_boot_reached(&self) -> bool {
+        self.end_of_boot.load(core::sync::atomic::Ordering::SeqCst)
+    }
+
+    /// Close VTL1's window of trusting VTL0. One-way.
+    pub(crate) fn signal_end_of_boot(&self) {
+        self.end_of_boot
+            .store(true, core::sync::atomic::Ordering::SeqCst);
     }
 
     /// Returns the physical frame range belonging to VTL1.
