+++ litebox_platform_lvbs/src/mshv/mod.rs
@@ -9,7 +9,6 @@ mod hvcall_vp;
 pub(crate) mod ringbuffer;
 pub mod vsm;
 pub mod vsm_intercept;
-pub mod vsm_mediation;
 pub mod vtl1_mem_layout;
 pub mod vtl_switch;
 
@@ -37,7 +36,7 @@ unsafe impl<const ALIGN: usize> VmapManager<ALIGN> for PrivilegedVmap {
         perms: PhysPageMapPermissions,
     ) -> Result<Self::MapInfo, PhysPointerError> {
         // SAFETY: callers uphold the raw mapping contract. This provider is used only for
-        // independently authorized HEKI patch and ring-buffer writes.
+        // writes whose destination the caller has independently authorized.
         unsafe { crate::platform_low().vmap_privileged(pages, perms) }
     }
 
@@ -72,8 +71,8 @@ unsafe impl<const ALIGN: usize> VmapManager<ALIGN> for PrivilegedVmap {
 type Vtl0PhysConstPtr<T, const ALIGN: usize> =
     litebox_common_linux::physical_pointers::PhysConstPtr<T, ALIGN, crate::Vmap>;
 
-/// Mutable VTL0 pointer reserved for validated HEKI text patching and the fixed-address log ring
-/// buffer. It bypasses ordinary protected-frame access checks and synchronization. Do not use it for other
+/// Mutable VTL0 pointer reserved for callers that have independently validated the destination.
+/// It bypasses ordinary protected-frame access checks and synchronization. Do not use it for other
 /// VTL0 destinations that could enable confused-deputy writes.
 type PrivilegedVtl0PhysMutPtr<T, const ALIGN: usize> =
     litebox_common_linux::physical_pointers::PhysMutPtr<T, ALIGN, PrivilegedVmap>;
